Refactor Timeseries CSV IO
